Resolved the meory leaks issue
The Doubly node structure had some meory leaks issues
due to the use of shared pointers for both it's previous
and next nodes which was creating a circular dependency
while trying to release the managed objects it own.
Replacing one of the node (the previous one) as weak
pointer removed that problem.