Added tree node structure for BST implementation
Also added explicit and std move_no_except for better check
and performance improvement. Modified the search logic in
doubly linked list so that the worst case complexity comes
down from O(N) to O(N/2) while finding or erasing an element.
This in turn effects/imroves the same for stack and queue
structures as well as those are built on top of this doubly
linked list data structure.