Introducing iterators for the AVL tree
Introducing various iterator classes for the AVL tree structure.
Although the functionalities may be very rudementary but the logic
behind how an iterator works and performs is the main objective.

As of now only forward iterator is implemented but with little
tweaks it can be extended for reverse and const iterators classes as
well (perhaps sometimes later)